Implemented: server side managing of DataTable. TODO: code clean up and marge to main.

+++ src/main/java/com/example/bootstraptabledemo/datatable/DataTableResponse.java
@@ -1,9 +1,9 @@
 package com.example.bootstraptabledemo.datatable;
 
 public interface DataTableResponse {
-    int getRecordsTotal();
+    Long getRecordsTotal();
 
-    int getRecordsFiltered();
+    Long getRecordsFiltered();
 
     int getDraw();
